ml_project/requirements.txt
click==7.1.2
pandas==1.1.5
pytest==6.2.3
model==0.6.0
scikit_learn==0.24.2
pandas_profiling==2.11.0
pydantic==1.8.2
dataclasses==0.6
PyYAML==5.4.1
